Rename builtin subscriber behavior
