+++ dds/src/implementation/dds_impl/mod.rs
@@ -3,6 +3,7 @@ pub mod any_data_writer_listener;
 pub mod any_topic_listener;
 pub mod behavior_builtin_data_reader_stateful;
 pub mod behavior_builtin_data_reader_stateless;
+pub mod behavior_builtin_subscriber;
 pub mod behavior_user_defined_data_reader;
 pub mod behavior_user_defined_topic;
 pub mod dds_data_reader;
@@ -12,7 +13,6 @@ pub mod dds_publisher;
 pub mod dds_subscriber;
 pub mod dds_topic;
 pub mod message_receiver;
-pub mod node_builtin_subscriber;
 pub mod node_domain_participant;
 pub mod node_user_defined_data_writer;
 pub mod node_user_defined_publisher;
